[BugFix][ThreadSafe] Add mutex protection for global_handler_map_ and session_handler_map_
- Add session_handler_mutex_ in DebugRouterCore
- Protect all access to global_handler_map_ with global_handler_mutex_
- Protect all access to session_handler_map_ with session_handler_mutex_
- Use std::shared_lock for read operations and std::unique_lock for write operations
- Fix potential data race issues when adding/removing handlers while messages are being processed
- Add comprehensive unit tests to verify thread safety and ensure the fix works correctly

+++ debug_router/native/net/websocket_task.cc
@@ -53,7 +53,7 @@ int GetErrorMessage() {
 
 static int readline(SOCKET sock, char *buf, size_t size) {
   char *out = buf;
-  while (out - buf < size) {
+  while (static_cast<size_t>(out - buf) < size) {
     int res = static_cast<int>(recv(sock, out, 1, 0));
     if (res == 1) {
       if (*out++ == '\n') {
@@ -386,8 +386,9 @@ bool WebSocketTask::do_read(std::string &msg) {
 
   msg.resize(payloadLen);
 
-  if (recv(socket_guard_->Get(), const_cast<char *>(msg.data()), payloadLen,
-           0) != payloadLen) {
+  if (static_cast<size_t>(recv(socket_guard_->Get(),
+                               const_cast<char *>(msg.data()), payloadLen,
+                               0)) != payloadLen) {
     LOGE("failed to read websocket message");
     onFailure("Failed to read websocket message, recv failed.",
               GetErrorMessage());
